test: isolate cluster restore from shared fixture (#28010)
